+++ src/syntax/ast/var_binding.rs
@@ -1,9 +1,11 @@
+use crate::lexic::token::Token;
+
 use super::Expression;
 
 #[derive(Debug)]
 pub struct Binding<'a> {
-    pub datatype: Option<String>,
-    pub identifier: Box<String>,
+    pub datatype: Option<&'a Token>,
+    pub identifier: &'a Token,
     pub expression: Expression<'a>,
     pub is_mutable: bool,
 }